requirements.txt
fastapi==0.115.2
openai==1.51.2
openpyxl ==3.1.5
pdfplumber==0.11.4
pydantic==2.9.2
python-multipart==0.0.12
python-pptx==1.0.2
uvicorn==0.32.0
XlsxWriter==3.2.0
pymupdf==1.24.13
reportlab==4.2.5
docx2pdf==0.1.8
pdf2docx==0.5.8
python-dotenv==1.0.1